Server/Controllers/GameBoardController.cs
using Microsoft.AspNetCore.Mvc;

namespace Server.Controllers
{
    /// <summary>
    /// Controller for GET-request giving info about game board
    /// </summary>
    /// <seealso cref="Microsoft.AspNetCore.Mvc.ControllerBase" />
    [Route("api/[controller]")]
    [ApiController]
    public class GameBoardController : ControllerBase
    {
        [HttpGet]
        [ProducesResponseType(200)]
        [ProducesResponseType(500)]
        public IActionResult GetData()
        {
            var gameBoard = GameBoardHolder.GetGameBoardInstance();
            return Ok(gameBoard);
        }
    }
}